+++ Octokit.GraphQL/Model/ProjectV2ItemFieldDateValue.cs
@@ -28,7 +28,7 @@ namespace Octokit.GraphQL.Model
         /// <summary>
         /// Identifies the primary key from the database.
         /// </summary>
-        public int? DatabaseId { get; }
+        public long? DatabaseId { get; }
 
         /// <summary>
         /// Date value for the field
