+++ Octokit.GraphQL/Model/PullRequest.cs
@@ -153,7 +153,7 @@ namespace Octokit.GraphQL.Model
         /// Identifies the primary key from the database.
         /// </summary>
         [Obsolete(@"`databaseId` will be removed because it does not support 64-bit signed integer identifiers. Use `fullDatabaseId` instead. Removal on 2024-07-01 UTC.")]
-        public int? DatabaseId { get; }
+        public long? DatabaseId { get; }
 
         /// <summary>
         /// The number of deletions in this pull request.
