+++ Octokit.GraphQL/Model/TeamDiscussionComment.cs
@@ -60,7 +60,7 @@ namespace Octokit.GraphQL.Model
         /// <summary>
         /// Identifies the primary key from the database.
         /// </summary>
-        public int? DatabaseId { get; }
+        public long? DatabaseId { get; }
 
         /// <summary>
         /// The discussion this comment is about.
